fix: performance issue caused by the render calls
mactop has high CPU usage. Running it with cpuprofiling shows that the
top CPU time is spent in the termui.Render calls (to be specific, in the
ioctl syscalls there).

I removed a bunch of unnecessary Render calls (there were some
duplicates) and replaced all the other calls with a throttled version: I
call Render in the main process but not more often than twice in the
single "update interval" period.

This should address issue #13.